[firmware] use a separate I2C bus for the right half
This should the likelihood of any issues with the connection to the
right half causing problems for the i2c devices in the left half.
This also makes it easier to independently reset the two halves.

This may also make it easier to detect the right half being connected or
disconnected if I put the pull-up resistors for the right I2C bus inside
the right half of the keyboard.  I should be able to detect the right
half being plugged in when SDA and SCL are pulled up in this case.