[firmware] janky workaround to TinyUSB race conditions
Add a check to tud_hid_ready() before attempting to send a HID report.
This check is racy, but is better than nothing.  The
tud_hid_keyboard_report() code crashes if called when the device is not
mounted.  It is unfortunately unsafe to call from anywhere other than
the TinyUSB task, and the TinyUSB task is unusable for other work on
FreeRTOS systems.