[firmware] add a device argument to the  Esp32UsbDevice constructor
This probably doesn't really matter since the esp32s2 and esp32s3 chips
only support a single USB device.  It seems like we might as well allow
specifying this as a parameter though, on the off chance that there is a
chip in the future with multiple USB devices.

Also make the constructor constexpr.