[usb] rename find_descriptor() to get_descriptor()
I initially called this "find" to help clarify that this is currently
implemented as a linear scan rather than some more efficient look-up.
However, this is somewhat of an implementation detail, and implemented
this way since the scan is quite likely one of the more efficient
implementations when there are only a very small number of descriptors.