[cad] change the numpad plate to be flat with the front corners
Put the plate exactly in line with the front corners, so it is flat all
the way to the front.

I think this looks a bit less interesting, but it will make it easier
to put a LEDs and optionally a small OLED display below the numberpad.
Having this be flat with the numpad will make it possible to put this
all on a single PCB with the numpad keys.