[kicad] swap LDO from AMS1117 to NXP167BMX330TBG
I originally used AMS1117 since it is available as a basic part from
JLCPCB.  However, it requires a 22uF tantalum capacitor on the output,
which is only available as extended part, negating the cost savings of
the basic part.

This switches to using the NXP167BMX330TBG chip instead, which has much
lower quiescent current draw, higher PSRR, and a smaller footprint.  It
is only rated for 700mA, compared to 1A for the AMS1117, but my board
generally shouldn't be drawing that much power through the 3.3V lines.
The maximum allowed current draw for USB 2.0 is only 500mA.  The LEDs
are also probably the highest current draw, and these are powered
directly from 5V rather than the LDO's 3.3V output.